refactor: re-structured foldering system, divded by entities/page features
