+++ package.json
@@ -15,6 +15,7 @@
     "next": "14.2.7",
     "react": "^18",
     "react-dom": "^18",
+    "react-hook-form": "^7.53.0",
     "react-intersection-observer": "^9.13.1"
   },
   "devDependencies": {
