+++ pages/forum/[articleId].jsx
@@ -3,15 +3,12 @@ import Link from "next/link";
 import Image from "next/image";
 import { useRouter } from "next/router";
 import { QueryClient, useQuery, dehydrate } from "@tanstack/react-query";
-import { formatDate, formatLikes } from "@/lib/utils";
 import { getArticleById } from "@/lib/api";
-import KebabMenu from "@/components/ui/KebabMenu";
-import ProfileImg from "@/components/ui/ProfileImg";
-import inactiveHeart from "@/public/assets/icons/ic_heart_inactive.svg";
-import CommentList from "@/components/CommentList";
-import returnIcon from "../../public/assets/icons/ic_arrow_return.svg";
+import ArticleDetail from "@/components/article/ArticleDetail";
+import CommentList from "@/components/comment/CommentList";
 import Msg from "@/components/ui/Msg";
 import Loader from "@/components/ui/Loader";
+import returnIcon from "@/public/assets/icons/ic_arrow_return.svg";
 import styles from "@/styles/pages/forum/articleId.module.scss";
 
 export async function getServerSideProps(context) {
@@ -56,34 +53,7 @@ export default function ArticleDetailPage() {
       </Head>
 
       <section className={styles["article-section"]}>
-        <article>
-          <div className={styles.top}>
-            <h3>{article.title}</h3>
-            <KebabMenu />
-          </div>
-          <div className={styles.middle}>
-            <div className={styles.user}>
-              <ProfileImg width="40px" src={article.image} />
-              <span className={styles.name}>
-                {article.writer?.nickname || "총명한 판다"}
-              </span>
-              <span>{formatDate(article.createdAt)}</span>
-            </div>
-
-            <div className={styles["vertical-line"]}></div>
-
-            <button className={styles.likes}>
-              <Image
-                src={inactiveHeart}
-                alt="heart icon"
-                width={32}
-                height={32}
-              />
-              <span>{formatLikes(article.likeCount)}</span>
-            </button>
-          </div>
-          <p className={styles.bottom}>{article.content}</p>
-        </article>
+        <ArticleDetail article={article} />
 
         <CommentList />
