+++ pages/forum/index.jsx
@@ -9,10 +9,10 @@ import Head from "next/head";
 import Link from "next/link";
 import { getArticleList } from "@/lib/api";
 import { PAGE_SIZE } from "@/var";
-import BestArticles from "@/components/BestArticles";
-import Button from "@/components/ui/Button";
-import ArticleList from "@/components/ArticleList";
+import BestArticles from "@/components/article/BestArticles";
+import ArticleList from "@/components/article/ArticleList";
 import SearchBar from "@/components/form/SearchBar";
+import Button from "@/components/ui/Button";
 import DropDown from "@/components/ui/DropDown";
 import Msg from "@/components/ui/Msg";
 import Loader from "@/components/ui/Loader";
